Move webview into sub-crate/workspace rlib
This works around conflicting indirect dependencies, thought the exact
mechanisms are unclear. Presumably, by moving the webview calls/imports
into an rlib type crate, this creates a totally independent chain of
dependencies which are not deduplicated on, whilst still being statically
linkable.
Introduce another library subcrate to make shared types/functions
accessible in both the main binary and webview rlib.
