+++ Cargo.toml
@@ -8,8 +8,8 @@ description = "A mod manager for the game Starsector"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-druid = { git = "https://github.com/atlanticaccent/druid", rev = "76c130799d5341d11880b14b95a214dacf537fc3" }
-druid-widget-nursery = { git = "https://github.com/atlanticaccent/druid-widget-nursery", rev = "79b5368c26b5badd570959622a15834580ac7931", features = ["material-icons"] }
+druid = { git = "https://github.com/atlanticaccent/druid", rev = "2a54eb360aa171db133816c369144714221e345b" }
+druid-widget-nursery = { git = "https://github.com/atlanticaccent/druid-widget-nursery", branch = "switch-to-druid-fork", features = ["material-icons"] }
 infer = "0.3.4"
 tokio = { version = "1.6.0", features = ["fs", "io-util", "rt-multi-thread", "process", "macros"] }
 serde = { version = "1.0.126", features = ["derive"] }
@@ -32,7 +32,7 @@ classfile-parser = "~0.3"
 zip = "^0.5"
 regex = "1.5"
 lazy_static = "1.4"
-rfd = "0.6.3"
+rfd = "0.8"
 strum = "0.23"
 strum_macros = "0.23"
 tap = "1.0.1"
@@ -40,7 +40,6 @@ rayon = "1.5"
 self_update = { version = "0.29.0", default-features = false, features = ["rustls"] }
 indexmap = "1.8.0"
 chrono = "0.4.19"
-wry = { git = "https://github.com/atlanticaccent/wry.git", branch = "events-fork" }
 clap = { version = "3.1.6", features = ["derive"] }
 interprocess = "1.1.1"
 bincode = "1.3.3"
@@ -48,15 +47,37 @@ rand = "0.8.5"
 base64 = "0.13.0"
 const_format = "0.2.22"
 
+[dependencies.webview-subsystem]
+path = "./webview-subsystem"
+
+[dependencies.webview-shared]
+path = "./webview-subsystem-shared"
+
 [patch.crates-io]
-cairo-rs = { git = "https://github.com/gtk-rs/gtk-rs-core.git", tag = "0.15.1" }
-cairo-sys-rs = { git = "https://github.com/gtk-rs/gtk-rs-core.git", tag = "0.15.1" }
-gdk-sys = { git = "https://github.com/gtk-rs/gtk3-rs.git", tag = "0.15.1" }
-# gtk = { git = "https://github.com/gtk-rs/gtk3-rs.git", tag = "0.15.1" }
-glib-sys = { git = "https://github.com/gtk-rs/gtk-rs-core.git", tag = "0.15.1" }
-# gtk-sys = { git = "https://github.com/gtk-rs/gtk3-rs.git", tag = "0.15.1" }
+gobject-sys = { git = "https://github.com/gtk-rs/gtk-rs-core", tag = "0.15.10" }
+atk-sys = { git = "https://github.com/gtk-rs/gtk3-rs", tag = "0.15.4" }
+cairo-rs = { git = "https://github.com/gtk-rs/gtk-rs-core", branch = "0.15" }
+cairo-sys-rs = { git = "https://github.com/gtk-rs/gtk-rs-core", branch = "0.15" }
+gdk = { git = "https://github.com/gtk-rs/gtk3-rs", tag = "0.15.4" }
+gdk-sys = { git = "https://github.com/gtk-rs/gtk3-rs", tag = "0.15.4" }
+gtk = { git = "https://github.com/gtk-rs/gtk3-rs", tag = "0.15.4" }
+glib = { git = "https://github.com/gtk-rs/gtk-rs-core", branch = "0.15" }
+glib-sys = { git = "https://github.com/gtk-rs/gtk-rs-core", branch = "0.15" }
+pangocairo = { git = "https://github.com/gtk-rs/gtk-rs-core", branch = "0.15" }
+pango = { git = "https://github.com/gtk-rs/gtk-rs-core", branch = "0.15" }
+gtk-sys = { git = "https://github.com/gtk-rs/gtk3-rs", tag = "0.15.4" }
+gio = { git = "https://github.com/gtk-rs/gtk-rs-core", branch = "0.15" }
+gio-sys = { git = "https://github.com/gtk-rs/gtk-rs-core", branch = "0.15" }
+gdk-pixbuf = { git = "https://github.com/gtk-rs/gtk-rs-core", branch = "0.15" }
+gdk-pixbuf-sys = { git = "https://github.com/gtk-rs/gtk-rs-core", branch = "0.15" }
 
 [package.metadata.bundle]
 name = "Starsector Mod Manager"
 id = "org.laird.starsector_mod_manager"
 osx_minimum_system_version = "10.10"
+
+[workspace]
+members = [
+  "webview-subsystem",
+  "webview-subsystem-shared"
+]
