+++ webview-subsystem-shared/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "webview-shared"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = { version = "1.0.126", features = ["derive"] }
+lazy_static = "1.4"
+directories = "3.0"
+const_format = "0.2.22"
+interprocess = "1.1.1"
